Correction du type de statistique

+++ Assets/Scripts/CharacterData.cs
@@ -31,7 +31,7 @@ public class CharacterData : ScriptableObject, ITargetable
     public float rageDamageMultiplier = 0.1f;
     public int baseReflex;
     public float baseMobility;
-    public float baseVitality;
+    public int baseVitality;
     public int basePower;
     public int baseStability;
     public int baseSagacity;
